Fixed SQL Problems / Finished Tables / Added User List / + Favicon
I fixed the SQL problems in code so that the edit pages actually change the SQL table that is being displayed. Added the rest of the 8 tables to also be displayed and changed. Decided to add a way to see and edit users as well, still working on the edit portion but is good enough for now. Finally added a favicon that is an emoji plate + fork & knife.